pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fundamental-engine"
version = "0.1.0"
description = "Point-in-time fundamental data engine for equity research (EDGAR + Bloomberg)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Research Team" }]

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.6.0",
    "tenacity>=8.2.0",
    "diskcache>=5.6.0",
    "pyarrow>=14.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
    "black>=23.9.0",
    "types-requests>=2.31.0",
]

[project.scripts]
edgar_pull   = "fundamental_engine.cli.main:edgar_pull"
bbg_ingest   = "fundamental_engine.cli.main:bbg_ingest"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
